requirements.txt
django
django-firebird
fdb
django-filter
django-widget-tweaks